Module lmdb_zero::error

Error values and types returned by LMDB and this wrapper.

Enums

Error

Error type returned by LMDB.

Constants

BAD_DBI

The specified DBI was changed unexpectedly

BAD_RSLOT

Invalid reuse of reader locktable slot

BAD_TXN

Transaction must abort, has a child, or is invalid

BAD_VALSIZE

Unsupported size of key/DB name/data, or wrong DUPFIXED size

CORRUPTED

Located page was wrong type

CURSOR_FULL

Cursor stack too deep - internal error

DBS_FULL

Environment maxdbs reached

INCOMPATIBLE

Operation and DB incompatible, or DB type changed. This can mean:

INVALID

File is not a valid LMDB file

KEYEXIST

key/data pair already exists

MAP_FULL

Environment mapsize reached

MAP_RESIZED

Database contents grew beyond environment mapsize

NOTFOUND

key/data pair not found (EOF)

PAGE_FULL

Page has not enough space - internal error

PAGE_NOTFOUND

Requested page not found - this usually indicates corruption

PANIC

Update of meta page failed or environment had fatal error

READERS_FULL

Environment maxreaders reached

TLS_FULL

Too many TLS keys in use - Windows only

TXN_FULL

Txn has too many dirty pages

VERSION_MISMATCH

Environment version mismatch

Traits

LmdbResultExt

Extension methods for LMDB results

Type Definitions

Result

Result type returned for all calls that can fail.