Struct lmdb_zero::WriteTransaction

pub struct WriteTransaction<'env>(_);

A read-write LMDB transaction.

In addition to all operations valid on ConstTransaction, it is also possible to perform writes to the underlying databases.

Ownership

WriteTransactions can be created with all three ownership modes (but owned mode is not useful).

Example — Shared mode

use std::sync::Arc;

let env = Arc::new(create_env());
let db = Arc::new(lmdb::Database::open(
  env.clone(), None, &lmdb::DatabaseOptions::defaults()).unwrap());

// Type and lifetime annotated explicitly for clarity
let txn: lmdb::WriteTransaction<'static> = lmdb::WriteTransaction::new(
  env.clone()).unwrap();

// Do stuff with `txn`...

txn.commit().unwrap();

Lifetime

All notes for ConstTransaction apply.

Implementations

impl<'env> WriteTransaction<'env>

pub fn new<E>(env: E) -> Result<Self> where
E: Into<NonSyncSupercow<'env, Environment>>,

Creates a new, read-write transaction in the given environment.

Note

A transaction and its cursors must only be used by a single thread (enforced by the rust compiler), and a thread may only have a single read-write transaction at a time (even if NOTLS is in use — trying to start two top-level read-write transactions on the same thread will deadlock).

pub fn child_tx<'a>(&'a mut self) -> Result<WriteTransaction<'a>> where
'env: 'a,

Opens a new, read-write transaction as a child transaction of the given parent. While the new transaction exists, no operations may be performed on the parent or any of its cursors. (These bindings are actually stricter, and do not permit cursors or other references into the parent to coexist with the child transaction.)

After this call, whether or not it succeeds, it is possible to call access() on the original transaction again one more time, since the Rust borrow rules guarantee the old accessor was destroyed by the caller already.

Note

A transaction and its cursors must only be used by a single thread (enforced by the rust compiler).

Example

let db = lmdb::Database::open(
  &env, None, &lmdb::DatabaseOptions::defaults()).unwrap();
let mut txn = lmdb::WriteTransaction::new(&env).unwrap();
let f = lmdb::put::Flags::empty();
{
  let mut access = txn.access();
  access.put(&db, "Germany", "Berlin", f).unwrap();
  access.put(&db, "Latvia", "Rīga", f).unwrap();
  access.put(&db, "France", "Paris", f).unwrap();
}

{
  // Open a child transaction and do some more reading and writing.
  let subtx = txn.child_tx().unwrap();
  let mut access = subtx.access();
  assert_eq!("Berlin", access.get::<str,str>(&db, "Germany").unwrap());
  access.put(&db, "Germany", "Frankfurt", f).unwrap();
  assert_eq!("Frankfurt", access.get::<str,str>(&db, "Germany").unwrap());
  // Don't commit --- let the child transaction abort (roll back)
}

{
  let mut access = txn.access();
  // Now we can do some more reading and writing on the original
  // transaction.
  // The effect of the aborted child transaction are not visible.
  access.put(&db, "United Kingdom", "London", f).unwrap();
  assert_eq!("Berlin", access.get::<str,str>(&db, "Germany").unwrap());
}

{
  // Another child.
  let subtx = txn.child_tx().unwrap();
  {
    let mut access = subtx.access();
    access.put(&db, "Spain", "Madrid", f).unwrap();
  }
  // Commit this one this time.
  subtx.commit().unwrap();
}

{
  // Now the changes from the child are visible to this transaction,
  // but still not outside it.
  let mut access = txn.access();
  assert_eq!("Madrid", access.get::<str,str>(&db, "Spain").unwrap());
}

txn.commit().unwrap();

pub fn commit(self) -> Result<()>

Commits this write transaction.

pub fn access(&self) -> WriteAccessor<'_>

Returns a read/write accessor on this transaction.

Panics

Panics if an accessor has already been obtained from this transaction and not yet dropped.