Module lmdb_zero::traits

Module containing public traits.

This exists as a separate module solely so that it can be wildcard imported where necessary.

Re-exports

pub use error;

pub use error::LmdbResultExt;

Traits

AsLmdbBytes

Translates a value into a byte slice to be stored in LMDB.

AssocCursor

Types of transaction references which can be used to renew StaleCursors into functional Cursors.

CreateCursor

Types of transaction references which can be used to construct Cursors.

FromLmdbBytes

Inverts AsLmdbBytes, producing a reference to a structure inside a byte array.

FromReservedLmdbBytes

Like FromLmdbBytes, but can be used with put_reserve() calls.

LmdbOrdKey

Trait describing a value which can be used as an LMDB key by having LMDB call into the value’s Ord implementation.

LmdbOrdKeyIfUnaligned

Marker trait for types where Unaligned<T> is LmdbOrdKey.

LmdbRaw

Marker trait indicating a value is to be stored in LMDB by simply copying it in.

LmdbRawIfUnaligned

Marker trait for types where Unaligned<T> is LmdbRaw.

TxExt

Extension trait for Rc and Arc that allows up-casting a reference to ReadTransaction or WriteTransaction to ConstTransaction.