Trait lmdb_zero::traits::CreateCursor

pub trait CreateCursor<'txn> {
    fn cursor<'db, DB>(&self, db: DB) -> Result<Cursor<'txn, 'db>>
    where
        DB: Into<Supercow<'db, Database<'db>>>;
}

Types of transaction references which can be used to construct Cursors.

In most cases this is simply used as an extension trait (see the examples on Cursor). However, it can also be used to abstract over things that can be used to create Cursors if so desired.

Implementations are provided for references to the three general transaction types, as well as Rc and Arc directly wrapping either concrete transaction type with a 'static lifetime.

Required methods

fn cursor<'db, DB>(&self, db: DB) -> Result<Cursor<'txn, 'db>> where
DB: Into<Supercow<'db, Database<'db>>>,

Create a cursor using self as the reference to the containing transaction and db as the database the cursor will read from and write into.