Trait lmdb_zero::traits::TxExt

pub trait TxExt {
    type Const;
    fn to_const(self) -> Self::Const;
}

Extension trait for Rc and Arc that allows up-casting a reference to ReadTransaction or WriteTransaction to ConstTransaction.

Associated Types

type Const

Required methods

fn to_const(self) -> Self::Const

Returns self as a handle to a ConstTransaction rather than the subtype that was passed in.

This is still a shared handle to the same transaction.